fixed warnings

+++ src/SystemUtils.cpp
@@ -17,7 +17,7 @@
 
 double SystemUtils::timestamp() {
     auto time = std::chrono::system_clock::now().time_since_epoch();
-    double t = std::chrono::duration_cast<std::chrono::seconds>(time).count();
+    double t = (double)std::chrono::duration_cast<std::chrono::seconds>(time).count();
     t += (double)std::chrono::duration_cast<std::chrono::nanoseconds>(time).count() / 1e+9;
     return t;
 }
